refactor(Store and ares table): Rename module to be area module
1. Rename to be area module.
2. All operation scripts in a module only
provde local type.
3. Index of moduel assemble global types and export
to store scope.

Todo: So far the module filed of global type depends on
string match. Try to work out a way to globally manage all module
fileds.
